requirements.txt
flask==3.1.2
ldap3==2.9.1
flask-cors==4.0.0
requests==2.32.5
playwright==1.40.0
python-dotenv==1.0.0
